server/requirements.txt
fastapi
uvicorn
httpx
google-cloud-firestore